backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
boto3
opensearch-py
requests
python-dotenv
